Preactivity JAVA CLASS
Agregamos la funcionalidad de la primera vista de la app
